requirements.txt
h5py==3.9.0
matplotlib==3.8.0
numpy==1.23
opencv_python==4.8.0.76
openslide_python==1.2.0
pandas==1.4.2
Pillow==10.1.0
scikit_learn==1.3.1
scipy==1.10.0
tensorboardX==2.6.2.2
torch
torchvision
